third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGUseElement-svgdom-href1-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'href' property of the SVGUseElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS useElement.href.baseVal is "#MyRect"
PASS useElement.href.baseVal is "../custom/resources/rgb.svg#G"
PASS successfullyParsed is true

TEST COMPLETE
